Simulated data for use in a landmarking analysis

A combination of the data_repeat and data_outcomes data, modified to perform a landmark analysis. See the vignette associated with the Landmarking package (browseVignettes("Landmarking")) for more information about how this dataset was formed.

data_repeat_outcomes

Format

A dataset with 22888 rows and 17 columns.