Fit a landmarking model using a linear mixed effects (LME) model for the longitudinal submodel

This function performs the two-stage landmarking analysis. In the first stage the longitudinal submodel is fitted using the LME model and in the second stage the survival submodel is fitted. Integer specifying the number of bootstrap samples to take

fit_LME_landmark_model(
  data_long,
  x_L,
  x_hor,
  fixed_effects,
  random_effects,
  fixed_effects_time,
  random_effects_time,
  random_slope = TRUE,
  standardise_time = FALSE,
  patient_id,
  start_study_time,
  end_study_time,
  k,
  cross_validation_df,
  lme_control = nlme::lmeControl(),
  event_time,
  event_status,
  survival_submodel = c("standard_cox", "cause_specific", "fine_gray"),
  b
)

Arguments

data_long
Data frame containing repeat measurements data and time-to-event data in long format

x_L
Numeric specifying the landmark time

x_hor
Numeric specifying the horizon time

fixed_effects
Vector of character strings specifying the column names in data which correspond to the fixed effects

random_effects
Vector of character strings specifying the column names in data which correspond to the random effects

fixed_effects_time
Character string specifying the column name in data which contains the time at which the fixed effects were recorded

random_effects_time
Vector of character strings specifying the column names in data which contain the times at which repeat measures were recorded. This should either be length 1 or the same length as random_effects. In the latter case the order of elements must correspond to the order of elements in random_effects.

random_slope
Boolean indicating whether to include a random slope in the LME model

standardise_time
Boolean indicating whether to standarise the time variable by subtracting the mean and dividing by the standard deviation (see Details section for more information)

patient_id
Character string specifying the column name in data which contains the individual identifiers

start_study_time
Character string specifying the column name in data which indicates the time that an individual enters the study

end_study_time
Character string specifying the column name in data which indicates the time that an individual leaves the study

k
Integer specifying the number of folds for cross-validation

cross_validation_df
List of dataframes containing the cross-validation fold each individual is assigned to. Each dataframe in the list should be named according to the landmark time x_L that they correspond. Each dataframe should contain the columns patient_id and a column "cross_validation_number" which contains the cross-validation fold of the individual.

lme_control
Object created using nlme::lmeControl(), which will be passed to the control argument of the lme function

event_time
Character string specifying the column name in data which contains the event time

event_status
Character string specifying the column name in data which contains the event status (where 0=censoring, 1=event of interest, if there are competing events these are labelled 2 or above.)

survival_submodel
Character string specifying which survival submodel to use. Three options: the standard Cox model i.e. no competing risks ("standard_cox"), the cause-specific regression model ("cause_specific"), or the Fine Gray regression model ("fine_gray")

b
Integer specifying the number of bootstrap samples to take when calcluating standard error of c-index and Brier score

Value

List containing data, model_longitudinal, model_LME, model_LME_standardise_time, model_survival, and prediction_error.

data is a data frame that includes the LOCF values of the fixed_effects and estimates of the random_effects predicted from the LME model (see Details section). It also includes the predicted probability that the event of interest has occured by time x_L, labelled as "event_prediction". There is one row for each individual.

model_longitudinal indicates that the longitudinal submodel is LME.

model_LME contains the output from the lme function from package nlme. For a model using cross-validation, model_LME contains a list of outputs with each element in the list corresponds to a different cross-validation fold. prediction_error contains a list indicating the c-index and Brier score at time x_hor and their standard errors if parameter b is used. For more information on how the prediction error is calculated please see ?get_model_assessment which is the function used to do this within fit_LME_landmark_model.

model_LME_standardise_time contains a list of two objects mean_response_time and sd_response_time if the parameter standardise_time=TRUE is used. This is the mean and standard deviation use to normalise times when fitting the LME model.

model_survival contains the outputs from the survival submodel functions, including the estimated parameters of the model. For a model using cross-validation, model_survival will contain a list of outputs with each element in the list corresponding to a different cross-validation fold. model_survival contains the outputs from the function used to fit the survival submodel, including the estimated parameters of the model. For a model using cross-validation, model_survival contains a list of outputs with each element in the list corresponding to a different cross-validation fold. For more information on how the survival model is fitted please see ?fit_survival_model which is the function used to do this within fit_LME_landmark_model.

prediction_error contains a list indicating the c-index and Brier score at time x_hor and their standard errors if parameter b is used.

Details

There are two parts to fitting the landmark model: the longitudinal submodel and the survival submodel. This function id used to fit the LME model as the longitudinal submodel. This model will be described in more detail.

For an individual \(i\), the LME model can be written as

$$Y_i = X_i \beta + Z_i U_i + \epsilon_i$$

where

\(Y_i\) is the vector of outcomes at different time points for the individual

\(X_i\) is the matrix of covariates for the fixed effects at these time points

\(\beta\) is the vector of coefficients for the fixed effects

\(Z_i\) is the matrix of covariates for the random effects

\(U_i\) is the matrix of coefficients for the random effects

\(\epsilon_i\) is the error term, typically from N(0, \(\sigma\))

By using an LME model to fit repeat measures data we can allow measurements from the same individuals to be more similar than measurements from different individuals. This is done through the random intercept and/or random slope.

Extending this model to the case where there are multiple random effects, denoted \(k\), we have

$$Y_{ik} = X_{ik} \beta_k + Z_{ik} U_{ik} + \epsilon_{ik}$$

Using this model we can allow a covariance structure within the random effects term \(U_{ik}\), for example a sample from the multivariate normal (MVN) distribution \(MVN(0,\Sigma_u)\). This covariance structure means the value of one random effects variable informs about the value of the other random effects variables, leading to more accurate predictions and allowing there to be missing data in the random effects variables.

The function fit_LME_landmark_model uses this covariance structure for the random effects when fitting the LME model. To fit the LME model the function lme from the package nlme is used. The fixed effects are calculated as the LOCF for the variables fixed_effects at the landmark age x_L and the random effects are those stated in random_effects and at times random_effects_time. This model is used to predict the values of the random effects at the landmark time x_L.

It is important to distinguish between the validation set and the development set for fitting the LME model. The development set includes all the repeat measurements (including those after the landmark age x_L). Conversely, the validation set only includes the repeat measurements recorded up until and including the landmark age x_L.

There is an important consideration about fitting the linear mixed effects model. As the variable random_effects_time gets further from 0, the random effects coefficients get closer to 0. This causes computational issues as the elements in the covariance matrix of the random effects, \(\Sigma_u\), are constrained to be greater than 0. Using parameter standard_time=TRUE can prevent this issue by standardising the time variables to ensure that the random_effects_time values are not too close to 0.

The predictions of the random effects at the landmark age, in addition to the LOCF values for the fixed effects, are used as the covariates for the survival submodel. The values of these covariates can be viewed in the data frame data that are returned by this function.

For the survival submodel, there are three choices of model:

the standard Cox model, this is a wrapper function for coxph from the package survival

the cause-specific model, this is a wrapper function for CSC from package riskRegression

the Fine Gray model, this is a wrapper function for FGR from package riskRegression

The latter two models estimate the probability of the event of interest in the presence of competing events.

Author

Isobel Barrott isobel.barrott@gmail.com

Examples

if (FALSE) data(data_repeat_outcomes)
data_model_landmark_LME<-fit_LME_landmark_model(data_long=data_repeat_outcomes,
x_L=c(60,61),
x_hor=c(65,66),
k=10,
start_study_time="start_time",
end_study_time="event_time",
fixed_effects=c("ethnicity","smoking","diabetes"),
fixed_effects_time="response_time_sbp_stnd",
random_effects=c("sbp_stnd","tchdl_stnd"),
random_effects_time=c("response_time_sbp_stnd","response_time_tchdl_stnd"),
patient_id="id",
standardise_time = TRUE,
lme_control = nlme::lmeControl(maxIter=100,msMaxIter=100),
event_time="event_time",
event_status="event_status",
survival_submodel = "cause_specific")
#> [1] "Fitting longitudinal submodel, landmark age 60"
#> Warning: the condition has length > 1 and only the first element will be used
#> [1] "Complete, landmark age 60"
#> [1] "Fitting survival submodel, landmark age 60"
#> Warning: Loglik converged before variable  2 ; coefficient may be infinite. 
#> [1] "Complete, landmark age 60"
#> [1] "Fitting longitudinal submodel, landmark age 61"
#> Warning: the condition has length > 1 and only the first element will be used
#> [1] "Complete, landmark age 61"
#> [1] "Fitting survival submodel, landmark age 61"
#> Warning: Loglik converged before variable  2 ; coefficient may be infinite. 
#> [1] "Complete, landmark age 61"