Fit a landmarking model using a last observation carried forward (LOCF) model for the longitudinal submodel

This function performs the two-stage landmarking analysis. In the first stage, the longitudinal submodel is fitted using the LOCF model and in the second stage the survival submodel is fitted.

fit_LOCF_landmark_model(
  data_long,
  x_L,
  x_hor,
  covariates,
  covariates_time,
  start_study_time,
  end_study_time,
  k,
  cross_validation_df,
  patient_id,
  event_time,
  event_status,
  survival_submodel = c("standard_cox", "cause_specific", "fine_gray"),
  b
)

Arguments

data_long
Data frame containing repeat measurements data and time-to-event data in long format

x_L
Numeric specifying the landmark time

x_hor
Numeric specifying the horizon time

covariates
Vector of character strings specifying the column names in data which correspond to the covariates

covariates_time
Vector of character strings specifying the column names in data which contain the times at which covariates were recorded. This should either be length 1 or the same length as covariates. In the latter case the order of elements must correspond to the order of elements in covariates.

start_study_time
Character string specifying the column name in data which indicates the time that an individual enters the study

end_study_time
Character string specifying the column name in data which indicates the time that an individual leaves the study

k
Integer specifying the number of folds for cross-validation

cross_validation_df
List of dataframes containing the cross-validation fold each individual is assigned to. Each dataframe in the list should be named according to the landmark time x_L that they correspond. Each dataframe should contain the columns patient_id and a column "cross_validation_number" which contains the cross-validation fold of the individual.

patient_id
Character string specifying the column name in data which contains the individual identifiers

event_time
Character string specifying the column name in data which contains the event time

event_status
Character string specifying the column name in data which contains the event status (where 0=censoring, 1=event of interest, if there are competing events these are labelled 2 or above.)

survival_submodel
Character string specifying which survival submodel to use. Three options: the standard Cox model i.e. no competing risks ("standard_cox"), the cause-specific regression model ("cause_specific"), or the Fine Gray regression model ("fine_gray")

b
Integer specifying the number of bootstrap samples to take when calcluating standard error of c-index and Brier score

Value

List containing data, model_longitudinal, model_survival, and prediction_error.

data is a data frame that includes the covariates values calculated using the LOCF model (see Details section) and the predicted probability that the event of interest has occured by time x_L, labelled as "event_prediction". There is one row for each individual.

model_longitudinal indicates that the longitudinal submodel is LOCF.

model_survival contains the outputs from the function used to fit the survival submodel, including the estimated parameters of the model. For a model using cross-validation, model_survival contains a list of outputs with each element in the list corresponding to a different cross-validation fold. For more information on how the survival model is fitted please see ?fit_survival_model which is a function used within fit_LOCF_landmark_model.

prediction_error contains a list indicating the c-index and Brier score at time x_hor and their standard errors if parameter b is used. For more information on how the prediction error is calculated please see ?get_model_assessment which is the function used to do this within fit_LOCF_landmark_model.

Details

There are two parts to fitting the landmark model: the longitudinal submodel and the survival submodel.

For the longitudinal model, this function uses the most recent values of the covariates at the landmark age x_L. This is the LOCF model.

For the survival submodel, there are three choices of model:

the standard Cox model, this is a wrapper function for coxph from the package survival

the cause-specific model, this is a wrapper function for CSC from package riskRegression

the Fine Gray model, this is a wrapper function for FGR from package riskRegression

The latter two models estimate the probability of the event of interest in the presence of competing events.

Author

Isobel Barrott isobel.barrott@gmail.com

Examples

data(data_repeat_outcomes)
data_model_landmark_LOCF<-fit_LOCF_landmark_model(data_long=data_repeat_outcomes,
  x_L=c(60,61),
  x_hor=c(65,66),
  covariates=c("ethnicity","smoking","diabetes","sbp_stnd","tchdl_stnd"),
  covariates_time=c(rep("response_time_sbp_stnd",4),"response_time_tchdl_stnd"),
  start_study_time="start_time",
  end_study_time="event_time",
  k=10,
  patient_id="id",
  event_time="event_time",
  event_status="event_status",
  survival_submodel = "cause_specific")

#> [1] "Fitting longitudinal submodel, landmark age 60"
#> [1] "Complete, landmark age 60"
#> [1] "Fitting survival submodel, landmark age 60"

#> Warning: Loglik converged before variable  2 ; coefficient may be infinite. 

#> [1] "Complete, landmark age 60"
#> [1] "Fitting longitudinal submodel, landmark age 61"
#> [1] "Complete, landmark age 61"
#> [1] "Fitting survival submodel, landmark age 61"

#> Warning: Loglik converged before variable  2 ; coefficient may be infinite. 

#> [1] "Complete, landmark age 61"