Predict the risk of an event for a new individual using the landmark model

This function predicts the risk of an event for new data using the landmark model fitted by fit_LME_landmark_model or fit_LOCF_landmark_model. The event is that which has event_status equal to 1.

# S3 method for landmark
predict(object, x_L, x_hor, newdata, cv_fold = NA, ...)

Arguments

object
Object inheriting the class landmark, this should be the output from either fit_LME_landmark_model or fit_LOCF_landmark_model. It should contain a list of landmark models corresponding to different landmark times x_L.

x_L
Numeric specifying the landmark time. This indicates which landmark model in object to use.

x_hor
Numeric specifying the horizon time, The function assesses the risk of event before this time.

newdata
Data frame containing new data to return the risk prediction of the event of interest. The data should be in in long format and the columns must contain the covariates and time variables that are used to fit the model. For the LME model this the variables fixed_effects, random_effects, fixed_effects_time, and random_effects_time. For the LOCF model this is covariates and covariates_time. Unfortunately there seems to be a bug in other packages meaning that for categorical covariates, all the possible categories need to be contained in newdata.

cv_fold
If cross validation is used to fit fit_LME_landmark_model or fit_LOCF_landmark_model, then the cross validation fold to use when making risk predictions needs to be specified.

Value

Data frame newdata updated to contained a new column event_prediction