Select individuals in a dataset with a last observation carried forward (LOCF) at a landmark time

To fit the LOCF model, all individuals must have at least one non-NA entry by age x_L for all covariates. This function selects these individuals and removes the other rows.

return_ids_with_LOCF(data, patient_id, x_L, covariates, covariates_time)

Arguments

data
Data frame with repeated measurements in long format; each row corresponds to an assessment entry

patient_id
Character string specifying the column name in data which contains the individual identifiers

x_L
Numeric specifying the landmark time

covariates
Vector of character strings specifying the column names in data which correspond to the covariates

covariates_time
Vector of character strings specifying the column names in data which contain the times at which covariates were recorded. This should either be length 1 or the same length as covariates. In the latter case the order of elements must correspond to the order of elements in covariates.

Value

Data frame data updated to contain only rows of individuals with a LOCF at age x_L, other rows are removed

Details

Individuals have a LOCF if there is a non-NA entry for each of the covariates in covariates up until (and including) time x_L.

Author

Isobel Barrott isobel.barrott@gmail.com

Examples

data(data_repeat_outcomes)
data_repeat_outcomes<-return_ids_with_LOCF(data=data_repeat_outcomes,
  patient_id="id",
  covariates=c("ethnicity","smoking","diabetes","sbp_stnd","tchdl_stnd"),
  covariates_time=c(rep("response_time_sbp_stnd",4),"response_time_tchdl_stnd"),
  x_L=60)